+++ gui/src/app/core/Project/ProjectSerialization.ts
@@ -49,26 +49,6 @@ export const deserializeProjectFromLocalStorage = (
   }
 };
 
-export const serializeAsZip = async (
-  data: ProjectDataModel,
-): Promise<[Blob, string]> => {
-  const fileManifest = mapModelToFileManifest(data);
-  const folderName = replaceSpacesWithUnderscores(data.meta.title);
-  const zip = new JSZip();
-  const folder = zip.folder(folderName);
-  if (!folder) {
-    throw new Error("Error creating folder in zip file");
-  }
-  Object.entries(fileManifest).forEach(([name, content]) => {
-    if (content.trim() !== "") {
-      folder.file(name, content);
-    }
-  });
-  const zipBlob = await zip.generateAsync({ type: "blob" });
-
-  return [zipBlob, folderName];
-};
-
 export const parseFile = (fileBuffer: ArrayBuffer) => {
   const content = new TextDecoder().decode(fileBuffer);
   return content;
@@ -98,7 +78,7 @@ export const deserializeZipToFiles = async (zipBuffer: ArrayBuffer) => {
       const content = await file.async("arraybuffer");
       const decoded = new TextDecoder().decode(content);
       files[basename] = decoded;
-    } else {
+    } else if (!["run.R", "run.py"].includes(basename)) {
       throw new Error(
         `Unrecognized file in zip: ${file.name} (basename ${basename})`,
       );
